requirements.txt
alembic==1.0.0
Babel==2.6.0
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
dominate==2.3.1
elasticsearch==6.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
